+++ addon/components/image-layer.js
@@ -1,9 +1,15 @@
+import Ember from 'ember';
 import BaseLayer from 'ember-leaflet/components/base-layer';
+
+const {
+  deprecate,
+  isPresent
+} = Ember;
 
 export default BaseLayer.extend({
 
   leafletRequiredOptions: [
-    'imageUrl', 'bounds'
+    'url', 'bounds'
   ],
 
   leafletOptions: [
@@ -11,8 +17,26 @@ export default BaseLayer.extend({
   ],
 
   leafletProperties: [
-    'url', 'opacity'
+    'url', 'opacity', 'bounds'
   ],
+
+  init() {
+    let imageUrl = this.get('imageUrl');
+    if (isPresent(imageUrl)) {
+      deprecate(
+        'ember-leaflet/image-layer: The `imageUrl` attribute has been deprecated in favor of the observed attribute `url`.',
+        false,
+        {
+          id:    'ember-leaflet.image-layer.imageUrl',
+          until: '4.0.0',
+          url:   'https://github.com/miguelcobain/ember-leaflet/pull/143'
+        }
+      );
+      this.set('url', imageUrl);
+    }
+
+    this._super(...arguments);
+  },
 
   createLayer() {
     return this.L.imageOverlay(...this.get('requiredOptions'), this.get('options'));